moved components into a components folder
