--- src/TableRow/types.ts
@@ -1,7 +1,0 @@
-export interface TableRowProps {
-    index: number;
-    category: string;
-    headline: string;
-    clicks?: number;
-
-}
