+++ src/components/TableRow/types.ts
@@ -1,0 +1,6 @@
+export interface TableRowProps {
+  index: number;
+  category: string;
+  headline: string;
+  clicks?: number;
+}
